% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column.map.r
\name{ColumnMap}
\alias{ColumnMap}
\title{Create column map}
\usage{
ColumnMap(
  mesh_xcord = "mesh_xcord",
  mesh_ycord = "mesh_ycord",
  tag_no = "tag_no",
  indv_no = "indv_no",
  stem_xcord = "stem_xcord",
  stem_ycord = "stem_ycord",
  spc_japan = "spc_japan",
  gbh = "gbh",
  vine = "vine",
  note = "note",
  s_date = "s_date",
  dead = "dead",
  position_change = "position_change",
  year = "year"
)
}
\arguments{
\item{mesh_xcord}{x coordinate of grid mesh.}

\item{mesh_ycord}{y coordinate of grid mesh.}

\item{tag_no}{stem id, currently should be unique for all stems.}

\item{indv_no}{individual id.}

\item{stem_xcord}{x coordinate of stems.}

\item{stem_ycord}{y coordinate of stems.}

\item{spc_japan}{species name in Japanese.}

\item{gbh}{girth at breast height.}

\item{vine}{if TRUE, GBH measure includes vine(s).}

\item{note}{remarks for each stem in each observation.}

\item{s_date}{census date.}

\item{dead}{logical indicating dead stems.}

\item{year}{observation year.}
}
\description{
Create column name mapping between source data and Moni1000 format.
Names of arguments represent Moni1000 data names and argument values
represent column names of source data.
}
