% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/header.r
\name{Moni1000Header}
\alias{Moni1000Header}
\title{Create Monitoring 1000 file header data}
\usage{
Moni1000Header(
  DATA_CREATOR = NA,
  DATA_TITLE = NA,
  SITE_NAME = NA,
  PLOT_NAME = NA,
  PLOT_CODE = NA,
  PLOT_SIZE = NA,
  DATA_CREATED = format(Sys.Date(), "\%Y\%m\%d"),
  PARAMETER_DEFINITIONS = ParameterDefinitions()
)
}
\arguments{
\item{DATA_CREATOR}{name(s) of data creator(s).}

\item{DATA_TITLE}{data title.}

\item{SITE_NAME}{site name.}

\item{PLOT_NAME}{plot name.}

\item{PLOT_CODE}{plot code.}

\item{PLOT_SIZE}{description of plot size.}

\item{DATA_CREATED}{data creation date.}

\item{PARAMETER_DEFINITIONS}{a \code{ParameterDefinitions} object having parameter descriptions.}
}
\value{
Moni100Header object.
}
\description{
Create Monitoring 1000 file header data
}
