% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.census.r
\name{write_moni1000_tree_census}
\alias{write_moni1000_tree_census}
\title{Write census data into Moni1000 format file.}
\usage{
write_moni1000_tree_census(
  data,
  path,
  column_map,
  header,
  encoding = "UTF-8",
  stem_columns = DEFAULT_STEM_COLUMNS,
  census_columns = DEFAULT_CENSUS_COLUMNS
)
}
\arguments{
\item{data}{a data.frame having tree census data.}

\item{path}{path to write.}

\item{column_map}{a \code{\link{ColumnMap}} object.}

\item{header}{a \code{\link{Moni1000Header}} object.}

\item{encoding}{file encoding to write.}

\item{stem_columns}{column names for stem-level information.
These columns are placed on the left side of the resultant table.}

\item{census_columns}{column names for census-level information.
These columns are placed on the right side of the resultant table with
column names with census year.}
}
\description{
Write census data into Moni1000 format file.
}
